Update: move store call body into environment
